renames around fake-xpring-client
